refactor: reuse existing messages endpoint for ephemeral chat (closes #44)
