+++ server/routes/messages.js
@@ -3,11 +3,13 @@ import { Hono } from "hono";
 
 import { createMessageSchema } from "../../shared/schemas.js";
 import * as queries from "../db/queries.js";
-import { getDb } from "../utils/context.js";
+import { getDb, getUserId } from "../utils/context.js";
 import { streamMockResponse } from "../utils/mock-response.js";
 import { paramSchema } from "./conversations-schema.js";
 
 const validateId = zValidator("param", paramSchema);
+
+let firstMessageCounter = 0;
 
 export const messagesRouter = new Hono()
     .get("/", validateId, async (c) => {
@@ -30,9 +32,22 @@ export const messagesRouter = new Hono()
         const convId = c.req.valid("param").id;
         const data = c.req.valid("json");
 
+        let actualConvId = convId;
+        let conversationCreated = false;
+
+        // Handle special "__new__" ID - creates conversation first
+        if (convId === "__new__") {
+            const userId = getUserId(c);
+            firstMessageCounter++;
+            const title = `New Chat ${firstMessageCounter}`;
+            const conv = queries.createConversation(db, { title, userId });
+            actualConvId = conv.id;
+            conversationCreated = true;
+        }
+
         // 1. Create user message
         const userMsg = queries.createMessage(db, {
-            conversationId: convId,
+            conversationId: actualConvId,
             role: "user",
             mode: data.mode,
             content: data.content,
@@ -43,6 +58,14 @@ export const messagesRouter = new Hono()
         return new Response(
             new ReadableStream({
                 async start(controller) {
+                    // Send conversation first if created (so client can update URL)
+                    if (conversationCreated) {
+                        const newConv = queries.getConversationById(db, actualConvId);
+                        controller.enqueue(
+                            JSON.stringify({ type: "conversation", data: newConv }) + "\n",
+                        );
+                    }
+
                     // Send user message first
                     controller.enqueue(
                         JSON.stringify({ type: "userMessage", data: userMsg }) + "\n",
@@ -63,7 +86,7 @@ export const messagesRouter = new Hono()
                     } finally {
                         // Save assistant message to DB
                         const assistantMsg = queries.createMessage(db, {
-                            conversationId: convId,
+                            conversationId: actualConvId,
                             role: "assistant",
                             mode: data.mode,
                             content: null,
