+++ shared/schemas.js
@@ -29,11 +29,6 @@ const createMessageSchema = z.object({
     mode: z.enum(["player", "gm"]),
 });
 
-const firstMessageSchema = z.object({
-    prompt: z.string().min(1),
-    mode: z.enum(["player", "gm"]).optional(),
-});
-
 const updateUserSchema = z.object({
     name: z.string().min(1).max(100).optional(),
     mode: z.enum(["gm", "player"]).optional(),
@@ -54,7 +49,6 @@ export {
     conversationSchema,
     createConversationSchema,
     createMessageSchema,
-    firstMessageSchema,
     updateUserSchema,
     ensureTestUserSchema,
 };
